test: Consolidate OS RHC and display_name filtering tests to share fixtures
Group tests using setup_hosts_for_os_rhc_filtering into
TestOsRhcFiltering and tests using
setup_hosts_for_os_display_name_filtering into
TestOsDisplayNameFiltering, both in test_filter_hosts.py.
This narrows fixture scope from module to class so each
expensive fixture runs only once per class, and batch-creates
hosts in the display_name fixture for additional speedup.
